refactor: split domain layer into focused modules (format, number, gap)
- Extract display formatting functions (gaps, percentages, currencies) into dedicated `format.ts` module
- Extract French-locale number parsing and normalization into dedicated `number.ts` module
- Keep gap business rules (calculations and threshold classification) in `gap.ts
- Reorganize domain exports and tests by concern for better maintainability
