+++ packages/app/src/modules/domain/__tests__/format.test.ts
@@ -1,0 +1,70 @@
+import { describe, expect, it } from "vitest";
+
+import {
+	computePercentage,
+	computeProportion,
+	formatCurrency,
+	formatGap,
+	formatGapCompact,
+	formatTotal,
+} from "../shared/format";
+
+describe("formatGap", () => {
+	it("formats a gap with French decimal separator", () => {
+		expect(formatGap(5.3)).toBe("5,3 %");
+	});
+
+	it("returns dash for null", () => {
+		expect(formatGap(null)).toBe("-");
+	});
+});
+
+describe("formatGapCompact", () => {
+	it("formats without percent sign", () => {
+		expect(formatGapCompact(5.3)).toBe("5,3");
+	});
+
+	it("returns dash for null", () => {
+		expect(formatGapCompact(null)).toBe("-");
+	});
+});
+
+describe("computeProportion", () => {
+	it("computes percentage from count and total", () => {
+		expect(computeProportion("25", 100)).toBe("25,0 %");
+	});
+
+	it("returns dash when total is zero", () => {
+		expect(computeProportion("10", 0)).toBe("-");
+	});
+});
+
+describe("formatCurrency", () => {
+	it("formats a number with euro sign", () => {
+		expect(formatCurrency("1234.5")).toMatch(/1[\s\u202f]234,5 €/);
+	});
+
+	it("returns dash for undefined", () => {
+		expect(formatCurrency(undefined)).toBe("-");
+	});
+});
+
+describe("computePercentage", () => {
+	it("computes percentage from count and total", () => {
+		expect(computePercentage(25, 100)).toBe("25,0 %");
+	});
+
+	it("returns dash when total is zero", () => {
+		expect(computePercentage(10, 0)).toBe("-");
+	});
+});
+
+describe("formatTotal", () => {
+	it("formats value with unit", () => {
+		expect(formatTotal(1234.5, "€")).toMatch(/1[\s\u202f]234,5 €/);
+	});
+
+	it("returns dash for null", () => {
+		expect(formatTotal(null, "€")).toBe("-");
+	});
+});
